created plugins section and moved example to it
