+++ MapboxAndroidDemo/src/main/AndroidManifest.xml
@@ -125,7 +125,7 @@
                 android:value="com.mapbox.mapboxandroiddemo.MainActivity"/>
         </activity>
         <activity
-            android:name=".examples.mas.TrafficPluginActivity"
+            android:name=".examples.plugins.TrafficPluginActivity"
             android:label="@string/activity_mas_traffic_plugin_title">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
