+++ MapboxAndroidDemo/src/main/res/values/strings.xml
@@ -364,6 +364,7 @@
     <string name="minneapolis">Minneapolis</string>
     <string name="miami">Miami</string>
     <string name="extrusions_category">3D Extrusions</string>
+    <string name="plugins_category">Plugins</string>
 
     <!--Landing activity-->
     <string name="tv_create_account_button">Create account</string>
